Fix "void" bug
Famous bug in NSMBU and almost all Nintendo games on Wii U; the frame buffers are not cleared at the start of each frame and, instead, everything is drawn on top of the previous frame. This is fine if you can cover the entire frame buffer when redrawing, but isn't when you don't, leaving trails of previous frames.